+++ src/main/java/kitchenpos/order/tobe/delivery/application/DeliveryOrderService.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.delivery.application;
+package kitchenpos.order.tobe.delivery.application;
 
 public class DeliveryOrderService {
     // Not yet implemented
