+++ src/main/java/kitchenpos/order/tobe/eatin/application/EatInOrderService.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.eatin.application;
+package kitchenpos.order.tobe.eatin.application;
 
 public class EatInOrderService {
     // Not yet implemented
