+++ src/main/java/kitchenpos/order/tobe/takeout/application/TakeoutOrderService.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.takeout.application;
+package kitchenpos.order.tobe.takeout.application;
 
 public class TakeoutOrderService {
     // Not yet implemented
