+++ src/test/java/kitchenpos/Fixtures.java
@@ -3,11 +3,11 @@ package kitchenpos;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuGroup;
 import kitchenpos.menu.domain.MenuProduct;
-import kitchenpos.order.order.domain.Order;
-import kitchenpos.order.order.domain.OrderLineItem;
-import kitchenpos.order.order.domain.OrderStatus;
+import kitchenpos.order.domain.Order;
+import kitchenpos.order.domain.OrderLineItem;
+import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.ordertable.domain.OrderTable;
-import kitchenpos.order.order.domain.OrderType;
+import kitchenpos.order.domain.OrderType;
 import kitchenpos.product.domain.Product;
 
 import java.math.BigDecimal;
